create custom tab component
